pyproject.toml
[project]
name = "hse-sundai-hack-interview-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "langchain-ollama>=0.3.2",
    "langgraph>=0.4.1",
    "ollama>=0.4.8",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]
